+++ treipigmei-chess-ai/src/piece/Knight.java
@@ -4,10 +4,10 @@ import helpers.Flags.Colour;
 
 public class Knight extends Piece {
     
-    public Knight(Colour color, byte[] position) {
+    public Knight(Colour color, int[] position) {
         super(color, position);
-        x = new byte[]{-2, -2, -1, -1, 1, 1, 2, 2};
-    	y = new byte[]{-1, 1, -2, 2, -2, 2, -1, 2};
+        x = new int[]{-2, -2, -1, -1, 1, 1, 2, 2};
+    	y = new int[]{-1, 1, -2, 2, -2, 2, -1, 2};
     }
